+++ CHANGES.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Fixed
 
 - Avoid usage of binary.Append to keep Go version requirement to 1.21
+- Update libs x/net 0.26=>0.35, x/sys 0.21=>0.30, x/text 0.16=>0.22
 
 ## 0.7.1 - 2024-12-04